Add a new migration to make sure that fast forward column exists in projects table

+++ db/migrate/20150827121444_add_fast_forward_option_to_project.rb
@@ -12,6 +12,8 @@ class AddFastForwardOptionToProject < ActiveRecord::Migration
   end
 
   def down
-    remove_column(:projects, :merge_requests_ff_only_enabled)
+    if column_exists?(:projects, :merge_requests_ff_only_enabled)
+      remove_column(:projects, :merge_requests_ff_only_enabled)
+    end
   end
 end
